+++ lwip/lwip_cpp/MulticastLwIp.cpp
@@ -6,7 +6,7 @@
 
 namespace services
 {
-    void MulticastLwIp::JoinMulticastGroup(IPv4Address multicastAddress)
+    void MulticastLwIp::JoinMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv4Address multicastAddress)
     {
         ip4_addr_t groupAddress;
         IP4_ADDR(&groupAddress, multicastAddress[0], multicastAddress[1], multicastAddress[2], multicastAddress[3]);
@@ -14,7 +14,7 @@ namespace services
         assert(result == ERR_OK);
     }
 
-    void MulticastLwIp::LeaveMulticastGroup(IPv4Address multicastAddress)
+    void MulticastLwIp::LeaveMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv4Address multicastAddress)
     {
         ip4_addr_t groupAddress;
         IP4_ADDR(&groupAddress, multicastAddress[0], multicastAddress[1], multicastAddress[2], multicastAddress[3]);
@@ -22,7 +22,7 @@ namespace services
         assert(result == ERR_OK);
     }
 
-    void MulticastLwIp::JoinMulticastGroup(IPv6Address multicastAddress)
+    void MulticastLwIp::JoinMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv6Address multicastAddress)
     {
         ip6_addr_t groupAddress;
         IP6_ADDR(&groupAddress, PP_HTONL(multicastAddress[1] + (static_cast<uint32_t>(multicastAddress[0]) << 16)), PP_HTONL(multicastAddress[3] + (static_cast<uint32_t>(multicastAddress[2]) << 16)), PP_HTONL(multicastAddress[5] + (static_cast<uint32_t>(multicastAddress[4]) << 16)), PP_HTONL(multicastAddress[7] + (static_cast<uint32_t>(multicastAddress[6]) << 16)));
@@ -30,7 +30,7 @@ namespace services
         assert(result == ERR_OK);
     }
 
-    void MulticastLwIp::LeaveMulticastGroup(IPv6Address multicastAddress)
+    void MulticastLwIp::LeaveMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv6Address multicastAddress)
     {
         ip6_addr_t groupAddress;
         IP6_ADDR(&groupAddress, PP_HTONL(multicastAddress[1] + (static_cast<uint32_t>(multicastAddress[0]) << 16)), PP_HTONL(multicastAddress[3] + (static_cast<uint32_t>(multicastAddress[2]) << 16)), PP_HTONL(multicastAddress[5] + (static_cast<uint32_t>(multicastAddress[4]) << 16)), PP_HTONL(multicastAddress[7] + (static_cast<uint32_t>(multicastAddress[6]) << 16)));