+++ services/network/Multicast.hpp
@@ -1,10 +1,13 @@
 #ifndef SERVICES_MULTICAST_HPP
 #define SERVICES_MULTICAST_HPP
 
+#include "infra/util/SharedPtr.hpp"
 #include "services/network/Address.hpp"
 
 namespace services
 {
+    class DatagramExchange;
+
     class Multicast
     {
     protected:
@@ -14,10 +17,10 @@ namespace services
         virtual ~Multicast() = default;
 
     public:
-        virtual void JoinMulticastGroup(IPv4Address multicastAddress) = 0;
-        virtual void LeaveMulticastGroup(IPv4Address multicastAddress) = 0;
-        virtual void JoinMulticastGroup(IPv6Address multicastAddress) = 0;
-        virtual void LeaveMulticastGroup(IPv6Address multicastAddress) = 0;
+        virtual void JoinMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv4Address multicastAddress) = 0;
+        virtual void LeaveMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv4Address multicastAddress) = 0;
+        virtual void JoinMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv6Address multicastAddress) = 0;
+        virtual void LeaveMulticastGroup(infra::SharedPtr<DatagramExchange> datagramExchange, IPv6Address multicastAddress) = 0;
     };
 }